Repo clean up. Small modifications to readme and man

+++ Backup/getline.c
@@ -15,7 +15,7 @@
  * Return: number of chars read
  */
 
-int64_t my_getline(char **lineptr, int *n, int fd)
+int my_getline(char **lineptr, int *n, int fd)
 {
 	int i = 1;
 	size_t check_read;
